fixed issue where doubles from reduced matrix were converted to int

+++ ConsoleApp1/Preprocessing2.cs
@@ -262,7 +262,7 @@ namespace ConsoleApp1
 
                                     if (number != null)
                                     {
-                                        outputMatrix[m, k] = (int)number;
+                                        outputMatrix[m, k] = (double)number;
                                     }
 
                                 }
